Explicitly defined socket timeout in send()
If a response is not received by the send() function (SNMP is not fully configured on target device, using an incorrect community string, access restriction, etc), the socket does not appear to time-out. I explicitly set the socket time-out to a value of 2 seconds using a default argument in the send() function call.