Add support for endOfMibView and noSuchInstance
References #66 and #54

The values are now supported but this brings microscopic changes to the
API. I will keep this a *minor* change because the previous behaviour of
the API was incorrect, and code relying on that behaviour is hence
incorrect as well. The biggest impact this could have stems from the
fact that code which previously incorrectly raised a "NoSuchOID"
exception, now returns ``None`` (in the pythonic API) and either
``NoSuchInstance()`` or ``NoSuchObject()`` in the "raw" API.

So any code catching ``NoSuchOID`` might require changes.