Replace "END_OF_MIB_VIEW" with EndOfMibView
References #54

The old class "puresnmp.pdu.EndOfMibView" which was instantiated as a
END_OF_MIB_VIEW constant was used as a special sentinel value but was
incorrect. At the time that was written, the correct implementaion was
unknown to me. It has now been correctly implemented as a special SNMP
value with byte marker ``0x82`` and the old constant was replaced