Remove "SNMP" specifics from the "x690" layer
SNMP table OIDs are followed by an "entry" OID (normally the ID ``1``).
This detail is local to "SNMP" and is not related to x690. As the
"puresnmp.x690.util" module is only concerned about x690 and should be
oblivious to SNMP, this was removed. It is the job of the SNMP layer to
include the ``.1`` index in the calculation of base nodes.

This is now handled in the various ``table()`` functions.

References #68 and #69